feat: Do not call toSearchableArray when deleting models
When deleting models from search, we do not need to provide Elasticsearch with the full model data, just the "_id" field so that Elasticsearch knows which document to delete. Calling toSearchableArray on models that have been deleted from the database can cause issues if relations etc cannot be hydrated, and also creates an unnecessary database call. This commit makes use of a new optional argument in "elastic-scout-driver" DocumentFactory which indicates if we want to call toSearchableArray on the models or not.